+++ WebApiService/Data/LoginContext.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.EntityFrameworkCore;
+using WebApiService.Models;
+
+namespace WebApiService.Data
+{
+    public class LoginContext : DbContext
+    {
+        public LoginContext(DbContextOptions<LoginContext> optins)
+            : base(optins)
+        {
+        }
+
+        public DbSet<ConnectionModel> Connections { get; set; }
+    }
+}
