+++ WebApiService/WebApiService.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CommonBizModule\CommonBizModule.csproj" />
+  </ItemGroup>
+
 </Project>
